+++ snooty.toml
@@ -29,3 +29,4 @@ mongodb-client-encryption-version = "2.2.0-alpha.2"
 key-vault-long = "Key Vault collection"
 kms-long = "Key Management System"
 cmk-long = "Customer Master Key"
+dek-long = "Data Encryption Key"
