+++ source/fundamentals/encrypt-fields/qe.txt
@@ -46,3 +46,35 @@ You must pass an ``AutoEncrytionOptions`` object to the ``MongoClient`` construc
 to enable {+qe+} operations with the {+driver-short+}.
 
 .. include:: /includes/auto-encryption-options.rst
+
+.. example::
+
+   .. literalinclude:: /includes/qe/qe-options.js
+      :start-after: begin qe mongoclient
+      :end-before: end qe mongoclient
+      :language: javascript
+      :dedent:
+
+ClientEncryption Configuratioin
++++++++++++++++++++++++++++++++
+
+``ClientEncryption`` is a class exposed by ``mongodb-client-encryption`` that provides
+methods for interacting with your {+DEK+}s. You construct a ``ClientEncryption`` object
+by passing a ``MongoClient`` object to the ``ClientEncryption`` constructor, and
+configuration options.
+
+.. include:: /includes/client-encryption-options.rst
+
+.. example::
+
+   .. literalinclude:: /includes/qe/qe-options.js
+      :start-after: begin client encryption
+      :end-before: end client encryption
+      :language: javascript
+      :dedent:
+
+Further Reading
+---------------
+
+For thorough coverage of the core concepts, use-cases, examples, and
+reference, visit the main {+qe+} :ref:`documentation site <qe-manual-feature-qe>`.
